+++ test/libsolidity/smtCheckerTests/types/address_transfer_insufficient.sol
@@ -12,6 +12,6 @@ contract C
 // SMTEngine: all
 // SMTIgnoreCex: yes
 // ----
+// Warning 8656: (101-116): CHC: Insufficient funds happens here.
+// Warning 8656: (120-136): CHC: Insufficient funds happens here.
 // Warning 6328: (180-204): CHC: Assertion violation happens here.
-// Warning 1236: (101-116): BMC: Insufficient funds happens here.
-// Warning 1236: (120-136): BMC: Insufficient funds happens here.